requirement.txt
aiohttp
discord
aiocron
beautifulsoup4
pydantic
pytz
PyGithub
fire
metagpt